Rename syllabus to outline.
